app/requirements.txt
flask==2.0.1
werkzeug==2.0.1
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-flask==0.42b0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-distro[otlp]==0.42b0
protobuf>=3.12.0
opentelemetry-exporter-otlp==1.21.0